requirements.txt
# =============================================================================
# OPTIMIZED REQUIREMENTS FOR FASTER DOCKER BUILDS
# =============================================================================
# Dependencies are organized by build complexity and installation time
# Heavy packages are grouped together for better Docker layer caching

# =============================================================================
# TIER 1: HEAVY SCIENTIFIC COMPUTING (Install first - longest build time)
# =============================================================================
# These packages have native dependencies and take the longest to compile
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0

# =============================================================================
# TIER 2: PDF AND DOCUMENT PROCESSING (Medium-heavy)
# =============================================================================
PyMuPDF==1.24.10
pypdf2==3.0.1
pdf2image==1.17.0

# =============================================================================
# TIER 3: DATABASE AND CORE INFRASTRUCTURE
# =============================================================================
# Database dependencies (install early to avoid conflicts)
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
alembic==1.14.0
pg8000==1.31.2
redis==5.1.0

# =============================================================================
# TIER 4: WEB FRAMEWORK AND API
# =============================================================================
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.28.1

#Graphql api
strawberry-graphql==0.282.0

# Data Processing and Utilities
python-multipart==0.0.12

# =============================================================================
# TIER 5: AUTHENTICATION AND SECURITY
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
email-validator==2.2.0

# =============================================================================
# TIER 6: AI/ML AND VECTOR DATABASES (Heavy but fewer conflicts)
# =============================================================================
openai==1.104.2
tiktoken==0.8.0
google-generativeai==0.7.2
azure-cognitiveservices-vision-computervision==0.9.0
pinecone==7.3.0
langchain-pinecone==0.2.12
langchain-google-genai==2.0.0
langchain-core==0.3.76
langchain_community
pgvector

# =============================================================================
# TIER 7: OCR AND IMAGE PROCESSING
# =============================================================================
pytesseract==0.3.10
rapidfuzz==3.9.7

# =============================================================================
# TIER 8: DATA PROCESSING AND UTILITIES (Lightweight)
# =============================================================================
pyyaml==6.0.2
openpyxl==3.1.5
filetype==1.2.0
requests==2.32.3
python-dotenv==1.0.1
tabulate==0.9.0
python-dateutil==2.8.2
ujson==5.9.0
jsonschema==4.23.0
tenacity==9.0.0
psutil==6.0.0
prometheus-client==0.21.0
cachetools==5.5.0
memory-profiler==0.61.0
better-profanity==0.7.0
# Profanity-filter: Install separately with --no-deps due to ordered-set-stubs dependency issue
# Run: pip install profanity-filter==1.3.3 --no-deps
# Profanity-filter dependencies (runtime deps, already listed below)
cached-property==1.5.2
more-itertools==8.14.0
ordered-set==3.1.1
ruamel.yaml>=0.17.0
spacy>=3.7.0
poetry-version==0.1.5

# =============================================================================
# TIER 9: TASK QUEUE AND BACKGROUND PROCESSING
# =============================================================================
celery==5.4.0

# =============================================================================
# TIER 10: CLI AND UTILITIES
# =============================================================================
click==8.1.7
pathlib-mate==1.3.2
playwright==1.48.0

# =============================================================================
# TIER 11: MONITORING AND LOGGING
# =============================================================================
sentry-sdk==2.14.0
structlog==24.4.0
colorlog==6.8.2

# =============================================================================
# TIER 12: DEVELOPMENT AND TESTING (Optional - can be excluded in production)
# =============================================================================
# Uncomment these for development builds only
# pytest==8.3.3
# pytest-asyncio==0.24.0
# pytest-cov==5.0.0
# black==24.8.0
# flake8==7.1.1